Added details and comments And fixed negative like counter bug

+++ unlike.php
@@ -10,29 +10,35 @@ session_start();
         include_once './classes/db.php';
         try{
             $pdo = dbConnection::connect();
+
+            $stmt = $pdo->prepare("SELECT * from likes WHERE username=? AND feedpostid=?");
+            $stmt->execute([$userid,$postid]);
+            $likes = $stmt->fetchAll(PDO::FETCH_ASSOC);
+   
+            if(!count($likes)){
+                throw new PDOException;
+            }
+
             $stmt = $pdo->prepare("DELETE FROM likes WHERE username=? AND feedpostid=?");
             $stmt->execute([$userid,$postid]);
-        }catch(PDOException){
-            echo 'ongoing';
-        }
-        try{
             $stmt = $pdo->prepare("SELECT likes from feedposts where id=?");
             $stmt->execute([$postid]);
             $likes = $stmt->fetchAll(PDO::FETCH_ASSOC);
             $likes = $likes[0]['likes'];
             $stmt = $pdo->prepare("UPDATE feedposts SET likes=? WHERE id=?");
             $stmt->execute([$likes-1,$postid]);
-        }catch(PDOException){
-            echo 'on going';
-        }
-        try{
             $stmt = $pdo->prepare("SELECT count('feedpostid') from likes where feedpostid=?");
             $stmt->execute([$postid]);
             $result = $stmt->fetchAll(PDO::FETCH_ASSOC);
             print_r($result[0]["count('feedpostid')"]);
-        }catch(PDOException){
-            echo 'on going';
+        }catch(PDOException $e){
+            $stmt = $pdo->prepare("SELECT count('feedpostid') from likes where feedpostid=?");
+            $stmt->execute([$postid]);
+            $result = $stmt->fetchAll(PDO::FETCH_ASSOC);
+            print_r($result[0]["count('feedpostid')"]);
+            // echo $e;
         }
+
        
 
     }else{
